Fix batch-small-files-2 byte mismatches
Two fixes needed to match the ROM:

1. unk_020551B8: UnkStruct_020551B8.bgm should be int, not u16.
   Loads were using ldr + lsls/lsrs cast dance; stores were using
   strh. With int bgm, loads become plain ldr and stores become
   plain str, matching the original.

2. unk_02097B78 LegendaryCinematic_Main: the original passes
   data (not data->args) to the cinematic function, via a cast.
   The compiler keeps data in r0 through the call, only using r1
   for the args-dereferenced index lookup. Likely a source quirk
   where the cinematic function accepts either pointer type or
   treats data as opaque state.

+++ src/unk_020551B8.c
@@ -14,7 +14,7 @@ typedef struct {
     int state;
     int result;
     int effect;
-    u16 bgm;
+    int bgm;
 } UnkStruct_020551B8;
 
 static BOOL sub_020551B8(TaskManager *man) {
